requirements.txt
chardet==5.2.0
Pillow==10.1.0
PyPDF2==3.0.1
reportlab==4.0.8
